Feature/service.space

+++ src/application/ui/manager/SprintManagerMenu.java
@@ -1,6 +1,5 @@
 package application.ui.manager;
 
-import bussinesslayer.entity.report.ReportSprint;
 import bussinesslayer.entity.space.Sprint;
 import bussinesslayer.service.report.reportsprint.IReportSprintService;
 import bussinesslayer.service.report.reportsprint.ReportSprintService;
@@ -8,7 +7,6 @@ import bussinesslayer.service.sapce.sprint.ISprintService;
 import bussinesslayer.service.sapce.sprint.SprintService;
 
 import java.time.LocalDate;
-import java.time.LocalTime;
 
 import static application.utilities.InputUtil.*;
 import static application.utilities.OutputUtil.*;
@@ -65,33 +63,57 @@ public class SprintManagerMenu {
             }
         }
     }
-    private void createSprint() throws Exception {
-        String name = readString("Name: ");
-        String description = readString("Description: ");
-        LocalDate startDate = readLocalDate("Start date: ");
-        LocalDate endDate = readLocalDate("End date: ");
-        Sprint sprint = new Sprint(name, description, startDate, endDate, projectId);
-        sprintService.createSprint(sprint);
+    private void createSprint() {
+        try {
+            String name = readString("Name: ");
+            String description = readString("Description: ");
+            LocalDate startDate = readLocalDate("Start date: ");
+            LocalDate endDate = readLocalDate("End date: ");
+            Sprint sprint = new Sprint(name, description, startDate, endDate, projectId);
+            sprintService.create(sprint);
+        } catch (Exception e) {
+            printValueln(e.getMessage());
+        }
     }
-    private void updateSprint() throws Exception {
-        int sprintId = readInt("Enter sprint id: ");
-        Sprint sprint = sprintService.getById(sprintId);
-        sprint.setName(readString("Enter new sprint name: "));
-        sprint.setDescription(readString("Enter new sprint description: "));
-        sprint.setStartDate(readLocalDate("Enter new sprint start date: "));
-        sprint.setEndDate(readLocalDate("Enter new sprint end date: "));
-        sprintService.update(sprint);
+    private void updateSprint() {
+        try {
+            int sprintId = readInt("Enter sprint id: ");
+            Sprint sprint = sprintService.getById(sprintId);
+            if (sprint.getProjectId() == projectId) {
+                sprint.setName(readString("Enter new sprint name: "));
+                sprint.setDescription(readString("Enter new sprint description: "));
+                sprint.setStartDate(readLocalDate("Enter new sprint start date: "));
+                sprint.setEndDate(readLocalDate("Enter new sprint end date: "));
+                sprintService.update(sprint);
+            } else {
+                printValueln("Sprint is not in this project.");
+            }
+        } catch (Exception e) {
+            printValue(e.getMessage());
+        }
+
     }
-    private void deleteSprint() throws Exception {
-        int sprintId = readInt("Enter sprint id: ");
-        sprintService.delete(sprintId);
+    private void deleteSprint() {
+        try {
+            int sprintId = readInt("Enter sprint id: ");
+            if (sprintService.getById(sprintId).getProjectId() == projectId) {
+                sprintService.delete(sprintId);
+            } else {
+                printValueln("Sprint is not in this project.");
+            }
+        } catch (Exception e) {
+            printValue(e.getMessage());
+        }
     }
-    private void viewSprint() throws Exception {
-        sprintService.viewById(projectId);
+    private void viewSprint() {
+        try {
+            sprintService.getAllSprintProject(projectId);
+        } catch (Exception e) {
+            printValue(e.getMessage());
+        }
     }
     private void viewReport() throws Exception {
         int sprintId = readInt("Enter sprint id: ");
-        sprintService.viewReport(sprintId);
     }
     private void processMenuForTaskManager() throws Exception {
         int sprintId = readInt("Enter sprint id: ");
